dom/tests/mochitest/geolocation/test_manyCurrentSerial.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=482260
-->
<head>
  <title>Test for getCurrentPosition </title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=482260">Mozilla Bug 482260</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

resume_geolocationProvider(function() {
  force_prompt(true, test1);
});

var keepGoing = 10;

function successCallback(position) {
  if (keepGoing-- > 0) {
    setTimeout(function() {navigator.geolocation.getCurrentPosition(successCallback);}, 0);
    return; 
  }

  ok(1, "100 successful calls");
  SimpleTest.finish();
}

function test1() {
  navigator.geolocation.getCurrentPosition(successCallback);
}
</script>
</pre>
</body>
</html>
